refactor: 온보딩 이미지 캐러셀 컴포넌트 분리
